fix avatar controller and typo

+++ app/Http/Controllers/Profile/StoreAvatarController.php
@@ -19,7 +19,7 @@ class StoreAvatarController extends Controller
             'image' => [
                 'image',
                 'mimes:jpg,jpeg,png',
-                'max:20000'
+                'max:20000',
             ],
             'crop.height' => 'integer|required',
             'crop.width' => 'integer|required',
@@ -31,8 +31,10 @@ class StoreAvatarController extends Controller
         $path = Str::random(40) . ".webp";
         Storage::disk('avatars')->put($path, $image);
         NewProfilePhotoEvent::dispatch(Auth::user(), $path);
-        if (Storage::disk('avatars')->exists(Auth::user()->profile_photo_path)) {
-            Storage::disk('avatars')->delete(Auth::user()->profile_photo_path);
+        if (!is_null(Auth::user()->profile_photo_path)) {
+            if (Storage::disk('avatars')->exists(Auth::user()->profile_photo_path)) {
+                Storage::disk('avatars')->delete(Auth::user()->profile_photo_path);
+            }
         }
         $request->user()->update(['profile_photo_path' => $path]);
         return Redirect::route('settings.profile');
